+++ src/app/apm/apm.module.ts
@@ -1,15 +1,19 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { SharedModule } from '../shared/shared.module';
 
 import { ApmRoutingModule } from './apm-routing.module';
 import { ErrorComponent } from './error/error.component';
 import { PerformanceComponent } from './performance/performance.component';
+import { SearchModule } from '../component/search';
+import { DetailComponent } from './error/detail/detail.component';
+import { ViewComponent } from './performance/view/view.component'
 
 @NgModule({
   imports: [
-    CommonModule,
-    ApmRoutingModule
+    SharedModule,
+    ApmRoutingModule,
+    SearchModule
   ],
-  declarations: [ErrorComponent, PerformanceComponent]
+  declarations: [ErrorComponent, PerformanceComponent, DetailComponent, ViewComponent]
 })
 export class ApmModule { }