+++ src/app/apm/performance/view/view.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'perf-view',
+  templateUrl: './view.component.html',
+  styleUrls: ['./view.component.scss']
+})
+export class ViewComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}