+++ src/app/component/component.module.ts
@@ -7,7 +7,6 @@ import { MarkdownComponent } from './markdown/markdown.component';
 import { PaginationComponent } from './pagination/pagination.component';
 import { PopoverComponent } from './popover/popover.component';
 import { PopoverDirective } from './popover/popover.directive';
-import { SearchComponent } from './search/search.component';
 
 import { TableComponent } from './table/table.component';
 
@@ -23,7 +22,6 @@ import { TableComponent } from './table/table.component';
     MarkdownComponent,
     PopoverComponent,
     PopoverDirective,
-    SearchComponent,
     TableComponent,
   ],
   exports: [
@@ -31,7 +29,6 @@ import { TableComponent } from './table/table.component';
     MarkdownComponent,
     PopoverComponent,
     PopoverDirective,
-    SearchComponent,
     TableComponent
   ]
 })