+++ src/app/component/index.ts
@@ -4,3 +4,4 @@ export * from './amap/index';
 export * from './loading/index';
 export * from './button/index';
 export * from './file-upload/index';
+export * from './search/index';