+++ src/app/pages/pages.routing..ts
@@ -6,6 +6,8 @@ import { ContactComponent } from './contact/contact.component';
 import { ServicesComponent } from './services/services.component';
 import { CollectionComponent } from './collection/collection.component';
 import { UserComponent } from './user/user.component';
+import { ProfileComponent } from './profile/profile.component';
+import { BlogComponent } from './blog/blog.component';
 
 const routes: Routes = [
   {
@@ -31,6 +33,14 @@ const routes: Routes = [
   {
     path: 'user',
     component: UserComponent
+  },
+  {
+    path: 'profile',
+    component: ProfileComponent
+  },
+  {
+    path: 'blog',
+    component: BlogComponent
   }
 ];